chore(settings): delete dead Connection files + slim list-models helper
Phase 1 of #889 cleanup: remove the source files that backed the Settings
Connection tab now that the tab itself is being deleted. The remaining
`readDefaultModelFromSettings` helper (still consumed by the
new-session model seed) is kept in a smaller, more accurately-named
`read-default-model.ts`. `modelPickerSlice` is renamed
`installerSlice` since the only fields left are the installer-corrupt
banner flag plus the boot-seeded default model id.
